Added new BSF method and node code
Deleted unecessary classes and updated main class to use some better code. Working on printing out the path to gold. Currently we know the shorts number of moves to get there but need to build a list of neighbors from each visited node so we can print the nodes taken to get to gold.